Added the swerve drive code
Added the draft code for the swerve drive modules and added some encoder canids in the constants.py file that need to be changed. Some parts of the swerve code need to be updated to the right values. These places are indicated by comments above them saying they need to be changed.